Fixed some logic, converted uint > unint32 for 5.1

+++ Public/Restart-ClusterNode.ps1
@@ -17,15 +17,21 @@ function Restart-ClusterNode {
 
         [bool]$awaitalive=$true,
 
-        [uint]$waittime=5,
+        [Uint32]$waittime=5,
 
-        [uint]$timebetweenpings=10,
+        [uint32]$timebetweenpings=10,
 
         [bool]$serialmode=$true
     )
     $awaitvalue = $null
     if($null -eq $inputobject){
-        Restart-Computer -ComputerName $node
+        $alive = Test-NetConnection -ComputerName $node
+        if($alive.pingsucceeded -eq $true){
+            Restart-Computer -ComputerName $node
+        }
+        else{
+            throw "Unable to resolve Node $($node)"
+        }
         write-host "Restart Command send to $($node)"
         if($awaitalive -eq $true){
             $count = 0
@@ -35,7 +41,7 @@ function Restart-ClusterNode {
                 $count++
                 Start-Sleep -Seconds $timebetweenpings
             }
-            write-output "Node $($node) is back online it took $($count * ($waittime * 60)) Minutes to complete."
+            write-output "Node $($node) is back online it took $(($count * $timebetweenpings) /60) Minutes to complete."
         }
     }
     else{
@@ -43,7 +49,13 @@ function Restart-ClusterNode {
             foreach($node in $inputobject){
                 $nodename = $null
                 $nodeName = $node.Name
-                Restart-Computer -ComputerName $nodename
+                $alive = Test-NetConnection -ComputerName $nodename
+                if($alive.pingsucceeded -eq $true){
+                    Restart-Computer -ComputerName $nodename
+                }
+                else{
+                    throw "Unable to resolve Node $($nodename)"
+                }
                 write-host "Restart Command send to $($nodename)"
                 if($awaitalive -eq $true){
                     $count = 0
@@ -53,24 +65,28 @@ function Restart-ClusterNode {
                         $count++
                         Start-Sleep -Seconds $timebetweenpings
                     } 
-                    write-output "Node $($nodename) is back online it took $($count * ($waittime * 60)) Minutes to complete."
+                    write-output "Node $($nodename) is back online it took $(($count * $timebetweenpings) /60) Minutes to complete."
                 }
             }
             Write-Output "Reboots all sent"
         }else{
-            restart-computer -ComputerName $inputobject.name
-            write-output "Restarts sent to $($inputobject.name)"
-            if($awaitalive -eq $true){
-                $inputobject.name | % -Parallel {
-                    $count = 0
-                    start-sleep -Seconds ($using:waittime * 60)
-                    while(($null -eq $using:awaitvalue) -or ($awaitvalue -ne "True")){
-                        $awaitvalue = Test-NetConnection -ComputerName $_
-                        $count++
-                        Start-Sleep -Seconds $timebetweenpings
-                    } 
-                    write-output "Node $($_) is back online it took $($count * ($waittime * 60)) Minutes to complete."
+            if(($PSEdition -eq "Core") -and ([UInt32]($PSVersionTable.PSVersion.ToString()[0]) -ge 7)){
+                restart-computer -ComputerName $inputobject.name
+                write-output "Restarts sent to $($inputobject.name)"
+                if($awaitalive -eq $true){
+                    $inputobject.name | % -Parallel {
+                        $count = 0
+                        start-sleep -Seconds ($using:waittime * 60)
+                        while(($null -eq $using:awaitvalue) -or ($awaitvalue -ne "True")){
+                            $awaitvalue = Test-NetConnection -ComputerName $_
+                            $count++
+                            Start-Sleep -Seconds $using:timebetweenpings
+                        } 
+                        write-output "Node $($_) is back online it took $(($count * $using:timebetweenpings) /60) Minutes to complete."
+                    }
                 }
+            }else{
+                throw "Parallel only works in PS7"
             }
         }
     }
